refactor(diff): align impl dir layout with cmd subcommand tree
`diff` is a top-level subcommand (pkg/cmd/diff/{db,detection}), but its
implementation lived under pkg/db/diff/{db,detection}, implying a "db diff"
hierarchy that does not match the CLI. Move it to pkg/diff/{db,detection}
to follow the pkg/cmd/X/Y <-> pkg/X/Y convention. CLI interface unchanged.
