+++ pkg/diff/db/testdata/fixtures/baseline/alma-errata/datasource.json
