+++ pkg/diff/db/testdata/fixtures/change-target/test-source/datasource.json
