+++ pkg/diff/db/testdata/fixtures/kb-baseline/microsoft-cvrf/datasource.json
