+++ pkg/diff/db/testdata/fixtures/kb-baseline/microsoft-cvrf/microsoftkb/KB5001000.json
