+++ pkg/diff/db/testdata/fixtures/kb-baseline/microsoft-cvrf/microsoftkb/KB5001111.json
