+++ pkg/diff/db/testdata/fixtures/kb-target-added/microsoft-cvrf/datasource.json
