+++ pkg/diff/db/testdata/fixtures/kb-target-added/microsoft-cvrf/microsoftkb/KB5001000.json
