+++ pkg/diff/db/testdata/fixtures/kb-target-changed/microsoft-cvrf/datasource.json
