+++ pkg/diff/db/testdata/fixtures/kb-target-changed/microsoft-cvrf/microsoftkb/KB5001000.json
