+++ pkg/diff/db/testdata/fixtures/kb-target-changed/microsoft-cvrf/microsoftkb/KB5001111.json
