+++ pkg/diff/db/testdata/fixtures/kb-target-same/microsoft-cvrf/datasource.json
