+++ pkg/diff/db/testdata/fixtures/kb-target-same/microsoft-cvrf/microsoftkb/KB5001000.json
