+++ pkg/diff/db/testdata/fixtures/target-added/alma-errata/data/ALSA/2019/ALSA-2019%3A3708.json
