+++ pkg/diff/db/testdata/fixtures/target-changed/alma-errata/data/ALSA/2024/ALSA-2024%3A0113.json
