+++ pkg/diff/db/testdata/fixtures/target-replaced/alma-errata/datasource.json
