+++ pkg/diff/db/testdata/fixtures/target-same/alma-errata/datasource.json
