+++ pkg/diff/detection/export_test.go
