+++ pkg/diff/detection/report.go
